+++ libinsdb/__init__.py
@@ -1,7 +1,6 @@
 # -*- encoding: utf-8 -*-
 
 from .objects import FormatSpecification, Entity, Quantity, DataFile, Release
-from .instrumentdb import InstrumentDatabase
 from .local import LocalInsDb, InstrumentDbFormatError
 from .remote import RemoteInsDb, InstrumentDbConnectionError